update to driver version 1.1.0
update main.c to changed sensor API,
initiaize driver with i2c address,
update signature for sensirion_i2c_hal_read/write

+++ sample-implementations/RaspberryPi_Pico/sensirion_i2c_hal.c
@@ -46,7 +46,7 @@
  * @param count   number of bytes to read from I2C and store in the buffer
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
+int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
     const bool nostop = true; // master retains control of the bus after the read
     return !i2c_read_blocking(i2c_default, address, data, count, nostop);
 }
@@ -63,7 +63,7 @@ int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
  * @returns 0 on success, error code otherwise
  */
 int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
-                               uint16_t count) {
+                               uint8_t count) {
     const bool nostop = true; // master retains control of the bus after the write
     return !i2c_write_blocking(i2c_default, address, data, count, nostop);
 }
